add volumeconfigs and switch to strategic merge talos patches
This moves the local-path-storage provisioner to a UserVolumeConfig and
sets up a UserVolumeConfig for ceph (when we are ready to test it).

You can not shrink the EPHEMERAL VolumeConfig on a running cluster,
that change can only be made by rebuilding the cluster. Per the Talos
docs, they recommend you have 100GB allocated for the EPHEMERAL
partition.

Ref: https://www.talos.dev/v1.10/introduction/system-requirements/#storage